refactor(caveTests): split cave tests in multiple files
