+++ package.json
@@ -43,8 +43,8 @@
     "start": "node app.js",
     "start-hot": "nodemon sails lift --silly",
     "debug": "node debug app.js",
-    "test": "NODE_ENV=test node ./node_modules/mocha/bin/mocha test/bootstrap.test.js test/integration/**/*.test.js",
-    "coverage": "NODE_ENV=test nyc --reporter=html --reporter=text mocha test/bootstrap.test.js test/integration/**/*.test.js",
+    "test": "NODE_ENV=test node ./node_modules/mocha/bin/mocha test/bootstrap.test.js \"test/integration/**/*.test.js\"",
+    "coverage": "NODE_ENV=test nyc --reporter=html --reporter=text mocha test/bootstrap.test.js \"test/integration/**/*.test.js\"",
     "coveralls": "NODE_ENV=test nyc report --reporter=text-lcov | coveralls",
     "format": "prettier --write \"**/*.+(js|json|md)\"",
     "snyk-protect": "snyk protect",
