+++ test/integration/3_routes/Cave/CAVE_PROPERTIES.js
@@ -1,0 +1,22 @@
+const CAVE_PROPERTIES = [
+  '@context',
+  '@id',
+  '@type',
+  'id',
+  'author',
+  'dateInscription',
+  'dateReviewed',
+  'depth',
+  'descriptions',
+  'documents',
+  'entrances',
+  'histories',
+  'isDeleted',
+  'isDiving',
+  'length',
+  'name',
+  'names',
+  'temperature',
+];
+
+module.exports = CAVE_PROPERTIES;
