+++ test/integration/3_routes/Cave/FAKE_DATA.js
@@ -1,0 +1,26 @@
+const caveCreationData = {
+  depth: 42,
+  descriptions: [
+    {
+      body: 'There are many rocks.',
+      language: 'eng',
+      title: 'Geology',
+    },
+    { body: "It's wet", language: 'eng', title: 'Hydrology' },
+    {
+      body: 'Very beautiful',
+      language: 'eng',
+      title: 'First gallery',
+    },
+  ],
+  documents: [1, 2],
+  isDiving: true,
+  latitude: 22.440979208779,
+  length: 530,
+  longitude: -83.980479240417,
+  massif: 1,
+  name: { text: 'La Cavité XYZ', language: 'eng' },
+  temperature: 8.36,
+};
+
+module.exports = caveCreationData;
